rewriting pr such that history is preserved in src/camera_calibration
